refactor: made sure that critical directories exists on start of app
