+++ .gitignore
@@ -1,5 +1,4 @@
 bin/
-node_modules/
 
 audiobook*
 assets/audios/*
