--- package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "tailwindcss": "^3.3.5"
-  }
-}
